+++ src/main/java/com/cadastro/apicadastro/services/PessoaService.java
@@ -5,13 +5,12 @@ import com.cadastro.apicadastro.dtos.MalaDiretaDTO;
 import com.cadastro.apicadastro.dtos.PessoaDTO;
 import com.cadastro.apicadastro.entities.Contato;
 import com.cadastro.apicadastro.entities.Pessoa;
+import com.cadastro.apicadastro.erros.excecoesTratadas.NotFoundException;
 import com.cadastro.apicadastro.mapper.PessoaMapper;
 import com.cadastro.apicadastro.repositories.PessoaRepository;
 import com.cadastro.apicadastro.requests.PessoaRegistroRequest;
 import com.cadastro.apicadastro.util.extensions.EnderecoExtensions;
 import jakarta.persistence.EntityNotFoundException;
-import jakarta.validation.Valid;
-import jakarta.validation.Validation;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -32,17 +31,16 @@ public class PessoaService {
     public PessoaDTO registraPessoa(PessoaRegistroRequest request) {
         Pessoa pessoa = PessoaMapper.INSTANCE.toPessoa(request);
 
-        for (Contato contato : request.getContatos()) {
-            contato.setPessoa(pessoa);
-        }
+        pessoaRepository.save(pessoa);
 
-        Pessoa pessoaSalva = pessoaRepository.save(pessoa);
-
-        return PessoaMapper.INSTANCE.toPessoaDTO(pessoaSalva);
+        return PessoaMapper.INSTANCE.toPessoaDTO(pessoa);
     }
 
     public Page<PessoaDTO> listaPessoas(Pageable pageable) {
-        Page<PessoaDTO> pessoasAtivas = pessoaRepository.findAllByAtivoTrue(pageable).map(PessoaMapper.INSTANCE::toPessoaDTO);
+        Page<PessoaDTO> pessoasAtivas = pessoaRepository
+                .findAllByAtivoTrue(pageable)
+                .map(PessoaMapper.INSTANCE::toPessoaDTO);
+
         return pessoasAtivas;
     }
 
@@ -51,13 +49,13 @@ public class PessoaService {
         if (pessoa != null) {
             return PessoaMapper.INSTANCE.toPessoaDTO(pessoa);
         } else {
-            return null; // LEMBRAR DE LANÇAR EXCEÇÃO
+            throw new NotFoundException("Pessoa não encontrada");
         }
     }
 
     @Transactional
     public PessoaDTO atualizaPessoa(Long id, AtualizaPessoaDTO atualiza) {
-        Pessoa pessoa = pessoaRepository.findById(id).orElseThrow();
+        Pessoa pessoa = pessoaRepository.findById(id).orElseThrow(() -> new NotFoundException("Pessoa não encontrada"));
         pessoa.atualizaPessoa(atualiza);
 
         return PessoaMapper.INSTANCE.toPessoaDTO(pessoa);
@@ -74,16 +72,18 @@ public class PessoaService {
             pessoaRepository.save(pessoa);
             return new PessoaDTO(pessoa);
         } else {
-            return null;
+            throw new NotFoundException("Pessoa não encontrada");
         }
     }
 
     @Transactional
     public PessoaDTO ativaPessoa(Long id) {
-        Pessoa pessoa = pessoaRepository.getReferenceById(id);
+        Pessoa pessoa = pessoaRepository.findById(id).orElseThrow(() -> new NotFoundException("Pessoa não encontrada"));
+
         pessoa.ativaPessoa();
 
         return PessoaMapper.INSTANCE.toPessoaDTO(pessoa);
+
     }
 
     @Transactional
@@ -96,13 +96,16 @@ public class PessoaService {
 
     @Transactional
     public void excluiPessoa(Long id) {
+        if (!pessoaRepository.existsById(id)) {
+            throw new NotFoundException("Pessoa não encontrada");
+        }
         pessoaRepository.deleteById(id);
     }
 
     public MalaDiretaDTO listaPessoaMalaDireta(Long id) {
         Optional<Pessoa> pessoaOptional = pessoaRepository.findById(id);
         if (pessoaOptional.isEmpty()) {
-            throw new EntityNotFoundException();
+            throw new NotFoundException("Pessoa não encontrada");
         }
 
         Pessoa pessoa = pessoaOptional.get();
